Group member list screen UI (#48)
* Add a screen to display the members of a group

This commit adds the UI which would show a list of members of the group.

* Indent

* Add a screen to display the members of a group

This commit adds the UI which would show a list of members of the group.

* Add a floatingActionButton

* Add profilePictureUrl and improve file naming

This commit adds a profilePictureUrl which would be used to display the profile picture and improves the naming of classes and files

* Add memberId and remove transparent_image for pubspec.yaml

This commit adds memberId and removes the transparent_image from pubspec.yaml. Also, an image for the placeholder has been added.